+++ NEWS.md
@@ -11,6 +11,7 @@
   * add addSearchOSM() in leafepp
   * add labels_es dataset to translate leafepp maps 
   * adaptation to osrm update
+  * add site by pkgdown
 
 ## EPP v0.3.2 (2021-08-03)