+++ NEWS.md
@@ -12,6 +12,7 @@
   * add labels_es dataset to translate leafepp maps 
   * adaptation to osrm update
   * add site by pkgdown
+  * add assertthat to check internet availability
 
 ## EPP v0.3.2 (2021-08-03)